refactor: move /books related code from domains/Notion
