--- src/domains/Notion/client.ts
@@ -1,5 +1,0 @@
-import { Client } from "@notionhq/client";
-
-export const NotionClient = new Client({
-	auth: process.env.NOTION_TOKEN,
-});
